+++ snuba/datasets/entities/storage_selectors/selector.py
@@ -1,39 +1,14 @@
-from abc import ABC, abstractmethod
-from typing import List, Type, cast
+from typing import List
 
+from snuba.datasets.entities.storage_selectors import QueryStorageSelector
 from snuba.datasets.storage import EntityStorageConnection, ReadableTableStorage
 from snuba.datasets.storages.storage_key import StorageKey
 from snuba.query.logical import Query
 from snuba.query.query_settings import QuerySettings
-from snuba.utils.registered_class import RegisteredClass
 
 
 class QueryStorageSelectorError(Exception):
     pass
-
-
-class QueryStorageSelector(ABC, metaclass=RegisteredClass):
-    """
-    The component provided by a dataset and used at the beginning of the
-    execution of a query to pick the storage query should be executed onto.
-    """
-
-    @classmethod
-    def config_key(cls) -> str:
-        return cls.__name__
-
-    @classmethod
-    def get_from_name(cls, name: str) -> Type["QueryStorageSelector"]:
-        return cast(Type["QueryStorageSelector"], cls.class_from_name(name))
-
-    @abstractmethod
-    def select_storage(
-        self,
-        query: Query,
-        query_settings: QuerySettings,
-        storage_connections: List[EntityStorageConnection],
-    ) -> EntityStorageConnection:
-        raise NotImplementedError
 
 
 class DefaultQueryStorageSelector(QueryStorageSelector):
